ex 7

+++ ex_3.py
@@ -1,6 +1,6 @@
 from exceptions import PeopleNotFoundException
 from all_peoples import AllPeoples
-from helpers import input_number
+from helpers import input_number, ask_confirmation
 
 
 def get_people_id():
@@ -8,14 +8,14 @@ def get_people_id():
 
 def delete_people_by_id(id):
   all_peoples = AllPeoples.build()
-
   all_peoples.delete_by_id(id)
 
 def delete_flow():
   people_id = get_people_id() # capturando id da pessoa
   try:
-    delete_people_by_id(people_id) # deletando ela do repositorio
-    print(f"Pessoa de id {people_id} apagada.")
+    if ask_confirmation():
+      delete_people_by_id(people_id) # deletando ela do repositorio
+      print(f"Pessoa de id {people_id} apagada.")
   except PeopleNotFoundException as exception: # tratando erro de pessoa nao encontrada
     print(str(exception))
